fix all 3 add report success
